:twisted_rightwards_arrows: Handle Coin and Laser, receive from the server (#32)
* :sparkles: handle coin receive from the server

* :bug: Update Makefile to include ncurses library for server binary

* :art: Rename authentication, game, laser, menu, and program files to include 'Graphic' suffix for clarity

* :bug: Fix order of flags in Makefile for server binary compilation

* :art: Add endian conversion macros for 64-bit integers in NetworksUtils.hpp

* :sparkles: Update payload size handling and endian conversion for player data passing to double for the player and the size_t

* :art: Refactor coin position handling to use double for precision and update related display methods

* :art: Refactor coin position handling to improve clarity and precision, add clearCoinPos method, and adjust payload size for COIN_POS

* :sparkles: Enhance game graphics by adding laser handling, updating display methods, and adjusting position calculations for coins and lasers

* :art: Refactor display methods for improved readability and consistency, update includes, and rename Program.hpp to ProgramGraphic.hpp

* :bug: Fix server error handling by ensuring menu switch occurs before exiting the application

+++ src/shared/include/Player.hpp
@@ -5,7 +5,7 @@
 typedef struct player {
     int id;
     char username[20];
-    float y_pos;
+    double y_pos;
     int coins_collected;
     bool is_dead;
     bool is_jetpack_on;
